Incoming call - code cleanup
