+++ CitiApps/SDK/communication-ui-library-ios-AzureCommunicationUICalling_1.1.0/AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/CallComposite.swift
@@ -10,7 +10,6 @@ import FluentUI
 import PIPKit
 import AzureCommunicationCalling
 
-var g_callAgent: CallAgent?
 public let callEndNotification = Notification.Name("acs_call_end")
 
 
@@ -86,11 +85,10 @@ public class CallComposite {
     /// - Parameter localOptions: LocalOptions used to set the user participants information for the call.
     ///                            This is data is not sent up to ACS.
     public func launch(remoteOptions: RemoteOptions,
-                       localOptions: LocalOptions? = nil, callAgent: CallAgent?) {
+                       localOptions: LocalOptions? = nil) {
         let callConfiguration = CallConfiguration(locator: remoteOptions.locator,
                                                   credential: remoteOptions.credential,
                                                   displayName: remoteOptions.displayName)
-        g_callAgent = callAgent
         launch(callConfiguration, localOptions: localOptions)
     }
