+++ CitiApps/SDK/communication-ui-library-ios-AzureCommunicationUICalling_1.1.0/AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Service/Calling/AzureCommunicationCalling/CallingSDKWrapper.swift
@@ -461,7 +461,6 @@ extension CallingSDKWrapper {
     }
 
     private func setupCallAgent() async throws {
-        callAgent = g_callAgent
         
         guard callAgent == nil else {
             logger.debug("Reusing call agent")
